build: Deal with git-checkout mtimes
Since this is a git checkout, in a fresh checkout the mtimes will not
reflect when each file was modified. Touch the configure script as soon
as the whole build process starts, in order to avoid spurious errors due
to files being newer than other files. Also, don't refresh old-configure
during the build process if we are in a git checkout.